Safari-609.1.11.1/LayoutTests/js/invalid-callframe-during-unwind-expected.txt

Test to ensure we have a valid callframe midway through unwinding

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testUnwind() threw exception threw successfully.
PASS successfullyParsed is true

TEST COMPLETE
